Move ViewSpectra component elements subfolder
