+++ app/packs/src/components/elements/ViewSpectra.js
